pyproject.toml
[tool.poetry]
name = "feed"
version = "1.1.2"
description = "arXiv feeds"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "3e19d4dd0e8d042a9ac6f9b0c1a9370a7d69aab0"}
feedgen = "==0.9.0"
flask = "^3"
Flask-SQLAlchemy = "^3"  
SQLAlchemy = "^2"
mysqlclient = "~=2.2.0"
werkzeug = "*"

[tool.poetry.dev-dependencies]

mypy = "*"
flake8 = "^3"
flask-testing = "^0.8"
pre-commit = "^2"
pydocstyle = "^5"
pylint = "^2"
pytest = "^7"
pytest-cov = "^2"
pytest-flask = "^1"
pytest-mock = "^3"